+++ controllers/tests/pick_up_heavy.py
@@ -29,9 +29,7 @@ if __name__ == "__main__":
     status = m.move_by(-1)
     assert status == ActionStatus.tipping, status
     status = m.move_by(1)
-    assert status == ActionStatus.tipping, status
-    status = m.move_by(-1)
-    assert status == ActionStatus.tipping, status
+    assert status == ActionStatus.collision, status
     status = m.move_by(-1)
     assert status == ActionStatus.success, status